+++ src/entities/reviews/index.ts
@@ -1,5 +1,5 @@
-import Review from './ui/review';
-export default Review;
+import MemoizedReview from './ui/review';
+export default MemoizedReview;
 
 export { reviewsSlice, reviewsActions } from './model/reviews-slice';
 export * from './model/selectors';
