+++ src/features/toolbar/index.ts
@@ -1,0 +1,2 @@
+import MemoizedToolbar from './ui/toolbar';
+export default MemoizedToolbar;
