+++ src/pages/favorites/favorites.tsx
@@ -1,5 +1,5 @@
 import Layout from '../../shared/layout';
-import Header from '../../widgets/header';
+import MemoizedHeader from '../../widgets/header';
 import MemoizedFooter from '../../widgets/footer';
 import SavedList from '../../widgets/saved-list';
 
@@ -7,7 +7,7 @@ const Favorites = (): JSX.Element => (
   <Layout
     wrapper="page"
     title="6 cities: favorites"
-    header={<Header />}
+    header={<MemoizedHeader />}
     content={
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
