+++ src/pages/main/main.tsx
@@ -4,7 +4,7 @@ import { useAppSelector, getPreviewOffersStatusObject, useActionCreators } from 
 import { getCurrentOffers, offersActions } from '../../entities/offers';
 
 import Layout from '../../shared/layout';
-import Header from '../../widgets/header';
+import MemoizedHeader from '../../widgets/header';
 import MemoizedFilter from '../../features/filter';
 import AllOffers from '../../widgets/all-offers';
 import BalloonLoader from '../../shared/ui/loader/balloon-loader';
@@ -23,7 +23,7 @@ const Main = (): JSX.Element => {
     <Layout
       wrapper="page page--gray page--main"
       title="6 cities"
-      header={<Header activeLogo />}
+      header={<MemoizedHeader activeLogo />}
       content={
         <main className={clsx('page__main page__main--index', { 'page__main--index-empty': !currentOffers.length })}>
           <h1 className="visually-hidden">Cities</h1>
