+++ src/widgets/header/index.ts
@@ -1,2 +1,2 @@
-import Header from './ui/header';
-export default Header;
+import MemoizedHeader from './ui/header';
+export default MemoizedHeader;
