+++ src/widgets/header/ui/header.tsx
@@ -1,11 +1,8 @@
-import { useCallback } from 'react';
-import { Link } from 'react-router-dom';
-import { AuthorizationStatus, AppRoute } from '../../../shared/enum';
-import { useAppSelector, getUserData } from '../../../shared/lib/redux';
-import { getAuthorizationStatus, useActionCreators } from '../../../shared/lib/redux';
-import { userActions } from '../../../entities/user';
+import { AppRoute } from '../../../shared/enum';
+import { memo } from 'react';
 
 import MemoizedLogo from '../../../shared/ui/logo';
+import MemoizedToolbar from '../../../features/toolbar';
 
 type HeaderProps = {
   withToolbar?: boolean;
@@ -13,14 +10,6 @@ type HeaderProps = {
 }
 
 const Header = ({ withToolbar = true, activeLogo = false }: HeaderProps): JSX.Element => {
-  const authorizationStatus = useAppSelector(getAuthorizationStatus);
-  const userData = useAppSelector(getUserData);
-  const { logOutAction } = useActionCreators(userActions);
-
-  const handleSignOutClick = useCallback(() => {
-    logOutAction();
-  }, [logOutAction]);
-
   if (!withToolbar) {
     return (
       <header className="header">
@@ -49,56 +38,13 @@ const Header = ({ withToolbar = true, activeLogo = false }: HeaderProps): JSX.El
               to={AppRoute.Root}
             />
           </div>
-          <nav className="header__nav">
-
-            <ul className="header__nav-list">
-
-              {authorizationStatus === AuthorizationStatus.Auth ?
-                (
-                  <>
-                    <li className="header__nav-item user">
-                      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
-                        <div className="header__avatar-wrapper user__avatar-wrapper"
-                          style={{
-                            backgroundImage: `url(${userData?.avatarUrl ?? '../img/avatar.svg'})`,
-                            borderRadius: '50%'
-                          }}
-                        >
-                        </div>
-                        <span className="header__user-name user__name">{userData?.email}</span>
-                        <span className="header__favorite-count">3</span>
-                      </Link>
-                    </li>
-                    <li className="header__nav-item">
-                      <Link
-                        className="header__nav-link"
-                        to="/"
-                        onClick={handleSignOutClick}
-                      >
-                        <span className="header__signout">Sign out</span>
-                      </Link>
-                    </li>
-                  </>
-                ) : (
-                  <li className="header__nav-item user">
-                    <Link
-                      className="header__nav-link header__nav-link--profile"
-                      to={AppRoute.Login}
-                    >
-                      <div className="header__avatar-wrapper user__avatar-wrapper">
-                      </div>
-                      <span className="header__login">Sign in</span>
-                    </Link>
-                  </li>
-                )}
-
-            </ul>
-
-          </nav>
+          {<MemoizedToolbar />}
         </div>
       </div>
     </header>
   );
 };
 
-export default Header;
+const MemoizedHeader = memo(Header);
+
+export default MemoizedHeader;
